Hand craft Python wrapper template and illustrate how we can use Cython array views and pure Python to maintain compatibility with old geosoft.gxapi API.
Added beneft of this approach is that we will no longer need a hard binary dependency on numpy. Note use of callback_free_data and print statements. The memory buffer does not seem to be copied once and is reused by numpy even. This approach is likely more efficient than the Boost Numpy approach.